bluetooth: support mbp16,2, fix lpss panic, disable read tx power
The 9001 patch was renamed to 8003 as we are running out of numbers for
sections of patches.

The new 9001 patch adds a quirk for mbp16,1/2/4 models that disables the
LE Read Min/Max Tx Power command, which is reported by the chip as a
supported command, but when Linux uses the command it fails. There is a
different patch fixing this on the bluetooth-next tree, but that doesn't
apply to stable.

9002 Adds the pcie id of the uart controller on the MBP16,2 to the
     intel-lpss-pci driver. The clock speed is corrected on the patch
     included.

9003 Fixes a kernel panic caused by an intel lpss acpi device, which
     doesn't seem to be used for anything on macOS (not much in ioreg
     about it).